backend/requirements.txt
flask 
flask-cors 
flask-jwt-extended 
pymongo 
bcrypt 
python-dotenv
fpdf